moved redux code into the component, moved epics into separate file
