+++ src/client/src/ui/spotTile/index.ts
@@ -4,3 +4,5 @@ export { default as PriceMovementIndicator } from './PriceMovementIndicator'
 export { default as TradeNotification } from './tradeNotification/'
 export { default as NotionalInput } from './notional/NotionalInput'
 export { default as SpotTileControls } from './spotTileControls/SpotTileControls'
+export { default as spotTileDataReducer } from './reducer'
+export { default as spotTileEpicsCreator } from './epics'
